Migrate old Instances to Internal, ledger type instances
