+++ src/Plutarch/Test/QuickCheck.hs
@@ -68,7 +68,7 @@ import Plutarch.Test.QuickCheck.Function (
   pattern PFn,
  )
 import Plutarch.Test.QuickCheck.Helpers (loudEval)
-import Plutarch.Test.QuickCheck.Instances (
+import Plutarch.Test.QuickCheck.Internal (
   PArbitrary (..),
   PCoArbitrary (..),
   TestableTerm (..),
@@ -379,7 +379,7 @@ instance
           Nothing -> property True
       Right (Right p', _, t) ->
         case h of
-          Just h' -> haskEquiv @e @( 'ByComplete) h' (TestableTerm p') Nil
+          Just h' -> haskEquiv @e @'ByComplete h' (TestableTerm p') Nil
           Nothing -> failWith $ "Haskell expected failure, but Plutarch succeed.\n" <> show t
 
 -- | @since 2.1.0
@@ -409,7 +409,7 @@ instance
   HaskEquiv 'OnBoth 'ByComplete h p '[]
   where
   haskEquiv h p _ =
-    haskEquiv @( 'OnPEq) @( 'ByComplete) h p Nil .&&. haskEquiv @( 'OnPData) @( 'ByComplete) h p Nil
+    haskEquiv @'OnPEq  @'ByComplete h p Nil .&&. haskEquiv @'OnPData  @'ByComplete h p Nil
 
 {- | Simplified version of `haskEquiv`. It will use arbitrary instead of
      asking custom generators.
